# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TodoCamisetas API",
        "description": "API REST para la gestión del inventario de TodoCamisetas. Incluye gestión de Clientes (Regulares/Preferenciales), Tallas y un sistema de cálculo dinámico de precios de Camisetas.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api/v1",
            "description": "Servidor Local Docker"
        }
    ],
    "paths": {
...
        "/shirts": {
            "get": {
                "tags": [
                    "Camisetas"
                ],
                "summary": "Listar todas las camisetas",
                "operationId": "getShirts",
                "responses": {
                    "200": {
                        "description": "Listado exitoso"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "post": {
                "tags": [
                    "Camisetas"
                ],
                "summary": "Crear una nueva camiseta",
                "operationId": "createShirt",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ShirtInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Camiseta creada"
                    },
                    "422": {
                        "description": "Errores de validación"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/clients/{id}/shirts": {
            "get": {
                "tags": [
                    "Camisetas"
                ],
                "summary": "Obtener las camisetas de un cliente específico",
                "description": "Busca un cliente por su ID y devuelve todas las camisetas que tiene asociadas, incluyendo sus tallas disponibles.",
                "operationId": "getShirtsByClient",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del cliente consultado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Shirt"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cliente no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Cliente no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/shirts/{id}": {
            "get": {
                "tags": [
                    "Camisetas"
                ],
                "summary": "Consultar precio final de camiseta",
                "description": "Calcula dinámicamente el precio final dependiendo de si el cliente es Regular o Preferencial, y si posee un porcentaje_oferta global.",
                "operationId": "getShirt",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "client_id",
                        "in": "query",
                        "description": "ID del cliente que consulta para calcular reglas de negocio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Camiseta con precio calculado"
                    },
                    "400": {
                        "description": "Falta parámetro client_id"
                    },
                    "404": {
                        "description": "Camiseta o cliente no encontrado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "put": {
                "tags": [
                    "Camisetas"
                ],
                "summary": "Actualizar una camiseta",
                "operationId": "updateShirt",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ShirtInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Camiseta actualizada"
                    },
                    "404": {
                        "description": "Camiseta no encontrada"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Camisetas"
                ],
                "summary": "Eliminar una camiseta",
                "operationId": "deleteShirt",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Camiseta eliminada"
                    },
                    "404": {
                        "description": "Camiseta no encontrada"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Shirt": {
                "title": "Shirt",
                "required": [
                    "cliente_id",
                    "titulo",
                    "club",
                    "pais",
                    "tipo",
                    "color",
                    "precio",
                    "codigo_producto"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "cliente_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "titulo": {
                        "type": "string",
                        "example": "Camiseta Local 2025"
                    },
                    "club": {
                        "type": "string",
                        "example": "Selección Chilena"
                    },
                    "pais": {
                        "type": "string",
                        "example": "Chile"
                    },
                    "tipo": {
                        "type": "string",
                        "example": "Local"
                    },
                    "color": {
                        "type": "string",
                        "example": "Rojo"
                    },
                    "precio": {
                        "type": "integer",
                        "example": 45000
                    },
                    "precio_oferta": {
                        "type": "integer",
                        "example": 35000,
                        "nullable": true
                    },
                    "detalles": {
                        "type": "string",
                        "example": "Edición especial",
                        "nullable": true
                    },
                    "codigo_producto": {
                        "type": "string",
                        "example": "SCL2025L"
                    }
                },
                "type": "object"
            },
            "ShirtInput": {
                "title": "Shirt Input",
                "required": [
                    "cliente_id",
                    "titulo",
                    "club",
                    "pais",
                    "tipo",
                    "color",
                    "precio",
                    "codigo_producto",
                    "sizes_ids"
                ],
                "properties": {
                    "cliente_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "titulo": {
                        "type": "string",
                        "example": "Camiseta Local 2025"
                    },
                    "club": {
                        "type": "string",
                        "example": "Selección Chilena"
                    },
                    "pais": {
                        "type": "string",
                        "example": "Chile"
                    },
                    "tipo": {
                        "type": "string",
                        "example": "Local"
                    },
                    "color": {
                        "type": "string",
                        "example": "Rojo"
                    },
                    "precio": {
                        "type": "integer",
                        "example": 45000
                    },
                    "precio_oferta": {
                        "type": "integer",
                        "example": 35000,
                        "nullable": true
                    },
                    "detalles": {
                        "type": "string",
                        "nullable": true
                    },
                    "codigo_producto": {
                        "type": "string",
                        "example": "SCL2025L"
                    },
                    "sizes_ids": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "example": [
                            1,
                            2,
                            3
                        ]
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}